Adapt chapters 2/5/8/9/11 incompatible tests to c6502's integer model
Modifies upstream test files in place to fit c6502's narrow
integer types: 1-byte int (-128..127), 2-byte long (-32768..32767),
4-byte long long (-2^31..2^31-1).

Chapter-by-chapter:
- chapter 2 (2 files): scale 2147483647 → 127; ~/-127 still
  exercises the unary-op-on-min/max-int boundary.
- chapter 5 (2 files): scale arithmetic / shift constants to fit
  in 1-byte int.
- chapter 8 (2 files): scale loop constants; preserve loop /
  iteration-statement semantics.
- chapter 9 (1 file): the 10M-iteration stack-arg leak test
  scaled to 8 iterations (still validates that sequential
  calls deallocate correctly).
- chapter 11 (28 files): substitute `long` → `long long` (since
  upstream's long = 8 bytes, c6502's = 2 bytes, but long long =
  4 bytes), scale literals from 8-byte to 4-byte ranges, and
  re-derive expected values for arithmetic / bitwise / shift /
  cast / switch / comparison tests. Test semantics survive:
  multi-byte arithmetic, common-type promotion, sign-/zero-
  extend, truncation, switch-on-wide-int, rewrite rules for
  constants beyond the immediate-byte range.

Also fixes a real compiler bug surfaced by the chapter-11 work:
`_tac_static_init_for` now applies modular truncation to the
declared integer type's width when laying down a static-storage
initializer (so e.g. `int i = 256ll;` lays down `dc.b $00`
rather than failing the byte-range check). Mirrors the runtime
cast lowering's byte-level semantics per C99 §6.3.1.3.

All 1245 unit tests + the 5 reactivated chapter harnesses pass.

+++ tests/chapter_11/valid/explicit_casts/sign_extend.c
@@ -1,22 +1,22 @@
-long sign_extend(int i, long expected) {
-    long extended = (long) i;
+long long sign_extend(int i, long long expected) {
+    long long extended = (long long) i;
     return (extended == expected);
 }
 
 
 int main(void) {
-    /* Converting a positive or negative int to a long preserves its value */
-    if (!sign_extend(10, 10l)) {
+    /* Converting a positive or negative int to a long long preserves its value */
+    if (!sign_extend(10, 10ll)) {
         return 1;
     }
 
-    if (!sign_extend(-10, -10l)) {
+    if (!sign_extend(-10, -10ll)) {
         return 2;
     }
 
-    /* sign-extend a constant to make sure we've implemented rewrite rule for movsx correctly */
-    long l = (long) 100;
-    if (l != 100l) {
+    /* sign-extend a constant to make sure we've implemented rewrite rule for sign-extend correctly */
+    long long l = (long long) 100;
+    if (l != 100ll) {
         return 3;
     }
     return 0;
